requirements.txt
requests>=2.31.0
beautifulsoup4>=4.9.3
flask>=3.0.0
aiohttp>=3.8.0
pydantic>=2.0.0
fastmcp>=0.1.0
rich>=10.0.0 